+++ sphinx/application/network/concepts/queries/concept-network-query-crypter/src/main/java/chat/sphinx/concept_network_query_crypter/model/SendSeedDto.kt
@@ -1,0 +1,11 @@
+package chat.sphinx.concept_network_query_crypter.model
+
+class SendSeedDto {
+    var seed: String? = null
+    var ssid: String? = null
+    var pass: String? = null
+    var lightningNodeIP: String? = null
+    var lightningNodePort: String? = null
+    var pubkey: String? = null
+    var network: String = "regtest"
+}