+++ sphinx/application/sphinx/src/main/java/chat/sphinx/di/NetworkModule.kt
@@ -10,6 +10,7 @@ import chat.sphinx.concept_network_client.NetworkClient
 import chat.sphinx.concept_network_client_cache.NetworkClientCache
 import chat.sphinx.concept_network_query_chat.NetworkQueryChat
 import chat.sphinx.concept_network_query_contact.NetworkQueryContact
+import chat.sphinx.concept_network_query_crypter.NetworkQueryCrypter
 import chat.sphinx.concept_network_query_invite.NetworkQueryInvite
 import chat.sphinx.concept_network_query_lightning.NetworkQueryLightning
 import chat.sphinx.concept_network_query_meme_server.NetworkQueryMemeServer
@@ -25,10 +26,12 @@ import chat.sphinx.concept_network_relay_call.NetworkRelayCall
 import chat.sphinx.concept_network_tor.TorManager
 import chat.sphinx.concept_relay.RelayDataHandler
 import chat.sphinx.concept_socket_io.SocketIOManager
+import chat.sphinx.concept_wallet.WalletDataHandler
 import chat.sphinx.feature_link_preview.LinkPreviewHandlerImpl
 import chat.sphinx.feature_network_client.NetworkClientImpl
 import chat.sphinx.feature_network_query_chat.NetworkQueryChatImpl
 import chat.sphinx.feature_network_query_contact.NetworkQueryContactImpl
+import chat.sphinx.feature_network_query_crypter.NetworkQueryCrypterImpl
 import chat.sphinx.feature_network_query_invite.NetworkQueryInviteImpl
 import chat.sphinx.feature_network_query_lightning.NetworkQueryLightningImpl
 import chat.sphinx.feature_network_query_meme_server.NetworkQueryMemeServerImpl
@@ -45,6 +48,7 @@ import chat.sphinx.feature_network_tor.TorManagerAndroid
 import chat.sphinx.feature_relay.RelayDataHandlerImpl
 import chat.sphinx.feature_socket_io.SocketIOManagerImpl
 import chat.sphinx.feature_sphinx_service.ApplicationServiceTracker
+import chat.sphinx.feature_wallet.WalletDataHandlerImpl
 import chat.sphinx.logger.SphinxLogger
 import chat.sphinx.meme_input_stream.MemeInputStreamHandlerImpl
 import chat.sphinx.wrapper_meme_server.AuthenticationToken
@@ -127,6 +131,27 @@ object NetworkModule {
         relayDataHandlerImpl: RelayDataHandlerImpl
     ): RelayDataHandler =
         relayDataHandlerImpl
+
+    @Provides
+    @Singleton
+    fun provideWalletDataHandlerImpl(
+        authenticationStorage: AuthenticationStorage,
+        authenticationCoreManager: AuthenticationCoreManager,
+        dispatchers: CoroutineDispatchers,
+        encryptionKeyHandler: EncryptionKeyHandler,
+    ): WalletDataHandlerImpl =
+        WalletDataHandlerImpl(
+            authenticationStorage,
+            authenticationCoreManager,
+            dispatchers,
+            encryptionKeyHandler
+        )
+
+    @Provides
+    fun provideWalletDataHandler(
+        walletDataHandlerImpl: WalletDataHandlerImpl
+    ): WalletDataHandler =
+        walletDataHandlerImpl
 
     @Provides
     @Singleton
@@ -401,4 +426,17 @@ object NetworkModule {
         networkQueryTransportKeyImpl: NetworkQueryRelayKeysImpl
     ): NetworkQueryRelayKeys =
         networkQueryTransportKeyImpl
+
+    @Provides
+    @Singleton
+    fun provideNetworkQueryCrypterImpl(
+        networkRelayCall: NetworkRelayCall
+    ): NetworkQueryCrypterImpl =
+        NetworkQueryCrypterImpl(networkRelayCall)
+
+    @Provides
+    fun provideNetworkQueryCrypter(
+        networkQueryCrypterImpl: NetworkQueryCrypterImpl
+    ): NetworkQueryCrypter =
+        networkQueryCrypterImpl
 }